Thor's solution for get_latest_balance_height TDD
